add authorization headers

+++ frontend/src/components/NoteDetails.js
@@ -1,13 +1,20 @@
 import { useNotesContext } from "../hooks/useNotesContext";
-import formatDistanceToNow from 'date-fns/formatDistanceToNow'
-
+import { useAuthContext } from "../hooks/useAuthContext";
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const NoteDetails = ({ note }) => {
+  const { user } = useAuthContext();
   const { dispatch } = useNotesContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/notes/" + note._id, {
-      method: "DELETE"
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
@@ -19,12 +26,16 @@ const NoteDetails = ({ note }) => {
   return (
     <div className="note-details">
       <h4>{note.title}</h4>
-      <h2><strong>{note.body}</strong></h2>
+      <h2>
+        <strong>{note.body}</strong>
+      </h2>
       <p>{formatDistanceToNow(new Date(note.createdAt), { addSuffix: true })}</p>
       <span className="material-symbols-outlined">edit</span>
-      <span className="material-symbols-outlined" onClick={handleDelete}>delete</span>
+      <span className="material-symbols-outlined" onClick={handleDelete}>
+        delete
+      </span>
     </div>
   );
 };
 
-export default NoteDetails 
+export default NoteDetails;
